Change ELF_Phdr.p_align for NetBSD/{amd64,arm} from 4k to the same size as a system standard ld(GNU linker).
        amd64: 0x200000 (2MiB)
        arm: 0x10000 (64KiB)

When p_align < PAGE_SIZE, NetBSD's ELF loader is not able to load a binary.
And the default PAGE_SIZE of NetBSD/arm kernel for many machines which have
armv6 or later CPU were changed from 4KiB to 8KiB since NetBSD 7.0.

I aligned to the same value as the default value of NetBSD standard binary
in amd64 too.

BTW, the standard value in NetBSD/i386 is the same 4KiB.